+++ src/App.jsx
@@ -1,16 +1,9 @@
 import React from "react";
-import { Routes, Route, Navigate } from "react-router-dom";
-import LoginPage from "./components/LoginPage";
-import AdminDashboard from "./components/AdminDashboard";
-import StudentDashboard from "./components/StudentDashboard";
+import { BrowserRouter } from "react-router-dom";
+import AppRoutes from "./routes/AppRoutes";
 
 export default function App() {
   return (
-    <Routes>
-      <Route path="/" element={<Navigate to="/login" />} />
-      <Route path="/login" element={<LoginPage />} />
-      <Route path="/admin" element={<AdminDashboard />} />
-      <Route path="/student" element={<StudentDashboard />} />
-    </Routes>
+      <AppRoutes />
   );
 }
